feat(one): add web.deploy cloudflare option for cloudflare worker deploys
