+++ packages/drizzle-to-zero/package.json
@@ -36,7 +36,7 @@
   },
   "devDependencies": {
     "@tamagui/build": "^1.120.1",
-    "vitest": "^2.1.1"
+    "vitest": "^2.1.8"
   },
   "publishConfig": {
     "access": "public"
