+++ packages/one/src/utils/ensureExists.ts
@@ -1,0 +1,5 @@
+export function ensureExists<T>(value: T | undefined | null): asserts value is T {
+  if (value === undefined || value === null) {
+    throw new Error(`Missing value.`)
+  }
+}
