+++ packages/one/types/serve-worker.d.ts
@@ -1,0 +1,3 @@
+import type { One } from './vite/types';
+export declare function serve(buildInfo: One.BuildInfo): Promise<import("hono").Hono<import("hono/types").BlankEnv, import("hono/types").BlankSchema, "/">>;
+//# sourceMappingURL=serve-worker.d.ts.map
