+++ packages/one/types/serve.d.ts
@@ -1,7 +1,6 @@
 import './polyfills-server';
-import type { Hono } from 'hono';
 import type { VXRNOptions } from 'vxrn';
-export declare function serve(args?: VXRNOptions['server']): Promise<void | Hono<import("hono/types").BlankEnv, import("hono/types").BlankSchema, "/"> | {
+export declare function serve(args?: VXRNOptions['server']): Promise<void | import("hono").Hono<import("hono/types").BlankEnv, import("hono/types").BlankSchema, "/"> | {
     handler: (req: Request, requestContext: import("hono/types").FetchEventLike) => Response | Promise<Response>;
     GET: (req: Request, requestContext: import("hono/types").FetchEventLike) => Response | Promise<Response>;
     POST: (req: Request, requestContext: import("hono/types").FetchEventLike) => Response | Promise<Response>;
