+++ packages/one/types/server/oneServe.d.ts
@@ -1,0 +1,5 @@
+import type { Hono } from 'hono';
+import type { VXRNOptions } from 'vxrn';
+import type { One } from '../vite/types';
+export declare function oneServe(oneOptions: One.PluginOptions, vxrnOptions: VXRNOptions, buildInfo: One.BuildInfo, app: Hono, serveStatic?: boolean): Promise<void>;
+//# sourceMappingURL=oneServe.d.ts.map
