+++ packages/one/types/server/setupBuildOptions.d.ts
@@ -1,0 +1,3 @@
+import type { One } from '../vite/types';
+export declare function setupBuildInfo(buildInfo: One.BuildInfo): void;
+//# sourceMappingURL=setupBuildOptions.d.ts.map
