+++ packages/one/types/utils/ensureExists.d.ts
@@ -1,0 +1,2 @@
+export declare function ensureExists<T>(value: T | undefined | null): asserts value is T;
+//# sourceMappingURL=ensureExists.d.ts.map
