+++ packages/vxrn/types/exports/serve.d.ts
@@ -1,10 +1,12 @@
+import type { Hono } from 'hono';
 import type { VXRNOptions, VXRNServePlatform } from '../types';
-export { createProdServer } from './createServer';
 export { loadEnv } from '../exports/loadEnv';
 export * from '../utils/getServerEntry';
+export { createProdServer } from './createServer';
 export declare const serve: (optionsIn: VXRNOptions & {
     platform?: VXRNServePlatform;
-}) => Promise<void | import("hono").Hono<import("hono/types").BlankEnv, import("hono/types").BlankSchema, "/"> | {
+    beforeStart?: (options: VXRNOptions, app: Hono) => void | Promise<void>;
+}) => Promise<void | Hono<import("hono/types").BlankEnv, import("hono/types").BlankSchema, "/"> | {
     handler: (req: Request, requestContext: import("hono/types").FetchEventLike) => Response | Promise<Response>;
     GET: (req: Request, requestContext: import("hono/types").FetchEventLike) => Response | Promise<Response>;
     POST: (req: Request, requestContext: import("hono/types").FetchEventLike) => Response | Promise<Response>;
