+++ tests/test-cloudflare/app/_layout.tsx
@@ -1,0 +1,10 @@
+import { LoadProgressBar, Slot } from 'one'
+
+export default function Layout() {
+  return (
+    <>
+      <LoadProgressBar />
+      <Slot />
+    </>
+  )
+}
