+++ tests/test-cloudflare/app/index+spa.tsx
@@ -1,0 +1,9 @@
+import { View } from 'react-native'
+
+export default function HomePage() {
+  return (
+    <>
+      <View style={{ width: 200, height: 200, backgroundColor: 'red' }} />
+    </>
+  )
+}
