+++ tests/test-cloudflare/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "test-cloudflare",
+  "version": "1.1.345-1734464750388",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "one dev",
+    "dev:clean": "one dev --clean",
+    "clean": "one clean",
+    "build:web": "one build",
+    "serve": "wrangler dev -c ./dist/wrangler.toml",
+    "test-failing": "vitest --run"
+  },
+  "dependencies": {
+    "one": "1.1.345-1734464750388",
+    "react-native": "0.74.5"
+  },
+  "devDependencies": {
+    "@cloudflare/vitest-pool-workers": "^0.5.38",
+    "@cloudflare/workers-types": "^4.20241216.0",
+    "vite": "^6.0.3",
+    "vitest": "^2.1.8",
+    "wrangler": "^3.97.0"
+  }
+}
