+++ tests/test-cloudflare/vite.config.ts
@@ -1,0 +1,12 @@
+import { one } from 'one/vite'
+import type { UserConfig } from 'vite'
+
+export default {
+  plugins: [
+    one({
+      web: {
+        deploy: 'cloudflare',
+      },
+    }),
+  ],
+} satisfies UserConfig
